Big API change in Cholesky module:  * rename Cholesky to LLT  * rename CholeskyWithoutSquareRoot to LDLT  * rename MatrixBase::cholesky() to llt()  * rename MatrixBase::choleskyNoSqrt() to ldlt()  * make {LLT,LDLT}::solve() API consistent with other modules
Note that we are going to keep a source compatibility untill the next beta release.
E.g., the "old" Cholesky* classes, etc are still available for some time.
To be clear, Eigen beta2 should be (hopefully) source compatible with beta1,
and so beta2 will contain all the deprecated API of beta1. Those features marked
as deprecated will be removed in beta3 (or in the final 2.0 if there is no beta 3 !).

Also includes various updated in sparse Cholesky.